Use uint*_t instead of char/short/int.

+++ c/src/lib/libcpu/powerpc/shared/include/byteorder.h
@@ -21,7 +21,7 @@
 
 #ifdef __GNUC__
 
-extern __inline__ unsigned ld_le16(volatile unsigned short *addr)
+extern __inline__ unsigned ld_le16(volatile uint16_t *addr)
 {
 	unsigned val;
 
@@ -29,12 +29,12 @@ extern __inline__ unsigned ld_le16(volatile unsigned short *addr)
 	return val;
 }
 
-extern __inline__ void st_le16(volatile unsigned short *addr, unsigned val)
+extern __inline__ void st_le16(volatile uint16_t *addr, unsigned val)
 {
 	__asm__ __volatile__ ("sthbrx %1,0,%2" : "=m" (*addr) : "r" (val), "r" (addr));
 }
 
-extern __inline__ unsigned ld_le32(volatile unsigned *addr)
+extern __inline__ unsigned ld_le32(volatile uint16_t *addr)
 {
 	unsigned val;
 
@@ -42,7 +42,7 @@ extern __inline__ unsigned ld_le32(volatile unsigned *addr)
 	return val;
 }
 
-extern __inline__ void st_le32(volatile unsigned *addr, unsigned val)
+extern __inline__ void st_le32(volatile uint32_t *addr, unsigned val)
 {
 	__asm__ __volatile__ ("stwbrx %1,0,%2" : "=m" (*addr) : "r" (val), "r" (addr));
 }
